pyproject.toml
[project]
name = "social-contagion-trends"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.68.0",
    "matplotlib>=3.10.8",
    "networkx>=3.6.1",
    "pandas>=3.0.1",
    "requests>=2.33.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
